5/28 기능 추가 및 수정 1. GamePlayUIController클래스(이름변경) -> GamePlayManager클래스에서 맵생성 함수를 스테이지버튼마다 onClick에 등록해줌 이 함수는 맵을 Instantiate하고 초기화해줌 2. GamePlayManager클래스에 CloseUI함수를 맵오브젝트에 OnPointerUp에 등록해주고 deltaPosition이 zero가 아닐때 실행하게 함 -> zero가 아니면 UI가 꺼지고 zero면 드레그 시에 UI가 그대로 남아있음 3. 유닛과 적에 HealthBar를 구현함 Image컴포넌트에 Fill로 체력이 줄어드는것을 구현함 Image이기 때문에 Canvas를 WorldSpace로 바꾸고 카메라를 LookAt() 시켜서 카메라를 항상 바라보게 만듬

+++ 3DTowerDefense/Assets/Scripts/TowerControl/ArcherTargetingTower.cs
@@ -38,7 +38,7 @@ namespace TowerControl
         }
 
         public override void TowerSetting(MeshFilter towerMeshFilter, int minDamage, int maxDamage, float range,
-            float delay, int health = 0)
+            float delay, float health = 0)
         {
             base.TowerSetting(towerMeshFilter, minDamage, maxDamage, range, delay, health);
