+++ 3DTowerDefense/Assets/Scripts/TowerControl/MageTargetingTower.cs
@@ -39,7 +39,7 @@ namespace TowerControl
         }
 
         public override void TowerSetting(MeshFilter towerMeshFilter, int minDamage, int maxDamage, float range,
-            float delay, int health = 0)
+            float delay, float health = 0)
         {
             base.TowerSetting(towerMeshFilter, minDamage, maxDamage, range, delay, health);
             CrystalPosInit();
@@ -52,7 +52,7 @@ namespace TowerControl
                 }
                 else
                 {
-                    //보라색 타워 전기처럼 
+                    //PurpleAttack 등록해줘야함
                 }
             }
             else
@@ -90,7 +90,7 @@ namespace TowerControl
 
         private void PurpleAttack()
         {
-            
+            //범위에 들어온 적 순서데로 발사체가 관통하며 공격 전기 이펙트 처럼 하면 좋겟는데 어케 만들지..
         }
     }
 }
