+++ 3DTowerDefense/Assets/Scripts/TowerControl/UnitTower.cs
@@ -44,7 +44,7 @@ namespace TowerControl
         *                                               Unity Event
         =========================================================================================================================================*/
 
-        protected abstract void UnitUpgrade(int minDamage, int maxDamage, float delay, int health);
+        protected abstract void UnitUpgrade(int minDamage, int maxDamage, float delay, float health);
         protected abstract void UnitSpawn(int i);
 
         protected override void Targeting()
@@ -95,7 +95,7 @@ namespace TowerControl
         }
 
         public override void TowerSetting(MeshFilter towerMeshFilter, int minDamage, int maxDamage, float range,
-            float delay, int health = 0)
+            float delay, float health = 0)
         {
             base.TowerSetting(towerMeshFilter, minDamage, maxDamage, range, delay, health);
             UnitUpgrade(minDamage, maxDamage, delay, health);
